+++ src/helpers/constants.py
@@ -131,6 +131,12 @@ class Endpoint(str, Enum):
     REPORTS_COMPLIANCE_TENANTS_TENANT_NAME = (
         '/reports/compliance/tenants/{tenant_name}'
     )
+    REPORTS_QUESTIONNAIRE_JOBS_JOB_ID = (
+        '/reports/questionnaire/jobs/{job_id}'
+    )
+    REPORTS_QUESTIONNAIRE_TENANTS_TENANT_NAME = (
+        '/reports/questionnaire/tenants/{tenant_name}'
+    )
     INTEGRATIONS_DEFECT_DOJO_ID_ACTIVATION = (
         '/integrations/defect-dojo/{id}/activation'
     )
@@ -713,6 +719,10 @@ class Env(EnvEnum):
 
     # Celery
     CELERY_BROKER_URL = 'SRE_CELERY_BROKER_URL', ('CAAS_CELERY_BROKER_URL',)
+    CELERY_RESULT_BACKEND = (
+        'SRE_CELERY_RESULT_BACKEND',
+        ('CAAS_CELERY_RESULT_BACKEND', 'CELERY_RESULT_BACKEND'),
+    )
     CELERY_TIMEZONE = 'SRE_CELERY_TIMEZONE', (), 'UTC'
     CELERY_TASK_COMPRESSION = 'SRE_CELERY_TASK_COMPRESSION', (), 'gzip'
     CELERY_WORKER_PREFETCH_MULTIPLIER = (
@@ -895,6 +905,16 @@ class Permission(str, Enum):
     REPORT_COMPLIANCE_DESCRIBE_JOB = 'report:get_job_compliance', False, True
     REPORT_COMPLIANCE_DESCRIBE_TENANT = (
         'report:get_tenant_compliance',
+        False,
+        True,
+    )
+    REPORT_QUESTIONNAIRE_DESCRIBE_JOB = (
+        'report:get_job_questionnaire',
+        False,
+        True,
+    )
+    REPORT_QUESTIONNAIRE_DESCRIBE_TENANT = (
+        'report:get_tenant_questionnaire',
         False,
         True,
     )